Manage snapshot DOM storage gracefully

+++ src/Codeception/Module/Percy/Persistence/DomStorage.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Codeception\Module\Percy\Persistence;
 
 use Codeception\Module\Percy\Exception\StorageException;
-use Ramsey\Uuid\Uuid;
 use Codeception\Module\Percy\Exchange\Action\Request\Snapshot;
+use Ramsey\Uuid\Uuid;
 
 class DomStorage
 {
@@ -16,11 +16,11 @@ class DomStorage
      * Save DOM snapshot to file
      *
      * @throws \Codeception\Module\Percy\Exception\StorageException
-     * @throws \Exception
-     * @param string $domString
-     * @return \Codeception\Module\Percy\Persistence\Dom
+     * @param \Codeception\Module\Percy\Exchange\Action\Request\Snapshot $snapshot
+     * @param string                                                     $domString
+     * @return \Codeception\Module\Percy\Exchange\Action\Request\Snapshot
      */
-    public function save(Snapshot $snapshot, string $domString): Dom
+    public function save(Snapshot $snapshot, string $domString): Snapshot
     {
         if (!function_exists('codecept_output_dir')) {
             throw new StorageException('`codecept_output_dir` function is not available!');
@@ -39,14 +39,20 @@ class DomStorage
 
         file_put_contents($filePath, $domString);
 
-        return Dom::from($filePath);
+        return $snapshot->withDomSnapshot(Dom::from($filePath));
     }
 
     /**
      * Clean snapshot directory
+     *
+     * @throws \tr33m4n\Utilities\Exception\AdapterException
      */
     public function clean(): void
     {
+        if (!config('percy')->get('cleanSnapshotStorage')) {
+            return;
+        }
+
         foreach (glob(codecept_output_dir(sprintf(self::OUTPUT_FILE_PATTERN, '*'))) ?: [] as $domFile) {
             unlink($domFile);
         }
